% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_indices.R
\name{all_indices}
\alias{all_indices}
\title{Calculate multiple vegetation indices.}
\usage{
all_indices(
  data,
  red_band_col,
  nir_band_col,
  redEdge_band_col,
  green_band_col,
  L_osavi = 0.16,
  L_savi = 0.5,
  L0 = 0.5
)
}
\arguments{
\item{data}{A data frame containing the spectral bands.}

\item{red_band_col}{A character string specifying the column name for the red band.}

\item{nir_band_col}{A character string specifying the column name for the NIR band.}

\item{redEdge_band_col}{A character string specifying the column name for the Red Edge band.}

\item{green_band_col}{A character string specifying the column name for the green band.}

\item{L_osavi}{A numeric value representing the L parameter for OSAVI.}

\item{L_savi}{A numeric value representing the L parameter for SAVI.}

\item{L0}{A numeric value representing the L parameter for MSAVI.}
}
\value{
A data frame with additional columns for each calculated vegetation index.
}
\description{
This function calculates various vegetation indices using the specified formulas and appends the results to the input data frame.
}
\examples{
df <- data.frame(
red_band = c(0.1, 0.2, 0.3, 0.4, 0.5),
nir_band = c(0.2, 0.3, 0.4, 0.5, 0.6),
redEdge_band = c(0.8, 0.9, 0.7, 1.0, 1.2),
green_band = c(0.4, 0.5, 0.6, 0.7, 0.8))
# Data frame
result_df <- all_indices(df, "red_band", "nir_band", "redEdge_band", "green_band")
# Results
result_df
}
