% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctvi.R
\name{ctvi}
\alias{ctvi}
\title{Calculate the Corrected Transformed Vegetation Index (CTVI) based on a previously calculated NDVI.}
\usage{
ctvi(red_band, nir_band)
}
\arguments{
\item{red_band}{A vector or matrix containing the red band values.}

\item{nir_band}{A vector or matrix containing the near-infrared (NIR) band values.}
}
\value{
A vector or matrix containing the calculated CTVI values.
}
\description{
The Corrected Transformed Vegetation Index (CTVI) is a vegetation index derived from NDVI. It is calculated as (NDVI + 0.5) / sqrt(abs(NDVI + 0.5)).
}
\examples{
red_band <- c(0.1, 0.2, 0.3, 0.4, 0.5)
nir_band <- c(0.2, 0.3, 0.4, 0.5, 0.6)

}
\references{
reference
Perry Jr, C. R., & Lautenschlager, L. F. (1984). Functional equivalence of spectral vegetation indices. Remote sensing of environment, 14(1-3), 169-182.
}
