% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvi.R
\name{dvi}
\alias{dvi}
\title{Calculate the Difference Vegetation Index (DVI).}
\usage{
dvi(red_band, nir_band)
}
\arguments{
\item{red_band}{A vector or matrix containing the values of the red band.}

\item{nir_band}{A vector or matrix containing the values of the NIR band.}
}
\value{
A vector or matrix containing the calculated DVI values.
}
\description{
The Difference Vegetation Index (DVI) is a simple vegetation index calculated as NIR - Red.
}
\examples{
red_band <- c(0.1, 0.2, 0.3, 0.4, 0.5)
nir_band <- c(0.2, 0.3, 0.4, 0.5, 0.6)
dvi(red_band, nir_band)
}
\references{
Richardson, A. J., & Everitt, J. H. (1992). Using spectral vegetation indices to estimate rangeland productivity. Geocarto International, 7(1), 63-69.
}
