% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_quanti.R
\name{evaluate_indices_rf}
\alias{evaluate_indices_rf}
\title{Evaluate Random Forest Models with Multiple Indices}
\usage{
evaluate_indices_rf(
  df,
  response_col,
  predictors,
  indices,
  train_percent,
  tune_grid
)
}
\arguments{
\item{df}{A data frame containing the input data.}

\item{response_col}{The name of the response variable.}

\item{predictors}{A character vector containing the names of predictor variables.}

\item{indices}{A character vector containing the names of indices to be tested.}

\item{train_percent}{The percentage of data to use for training.}

\item{tune_grid}{A data frame specifying the hyperparameter grid for tuning.}
}
\value{
None
}
\description{
This function evaluates Random Forest models for a given response variable, predictors, and a set of indices.
}
