% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gndvi.R
\name{gndvi}
\alias{gndvi}
\title{Calculate the Green Normalized Difference Vegetation Index (GNDVI).}
\usage{
gndvi(green_band, nir_band)
}
\arguments{
\item{green_band}{A vector or matrix containing the values of the green band.}

\item{nir_band}{A vector or matrix containing the values of the NIR band.}
}
\value{
A vector or matrix containing the calculated GNDVI values.
}
\description{
The Green Normalized Difference Vegetation Index (GNDVI) is a vegetation index that is calculated as (NIR - Green) / (NIR + Green).
}
\examples{
green_band <- c(0.1, 0.2, 0.3, 0.4, 0.5)
nir_band <- c(0.2, 0.3, 0.4, 0.5, 0.6)
gndvi(green_band, nir_band)
}
\references{
Gitelson, A. A., & Merzlyak, M. N. (1998). Remote sensing of chlorophyll concentration in higher plant leaves. Advances in Space Research, 22(5), 689-692.
}
