% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcari.R
\name{mcari}
\alias{mcari}
\title{Calculate the Modified Chlorophyll Absorption Ratio Index (MCARI).}
\usage{
mcari(red_band, redEdge_band, green_band)
}
\arguments{
\item{red_band}{A vector or matrix containing the values of the red band.}

\item{redEdge_band}{A vector or matrix containing the values of the Red Edge 1 band.}

\item{green_band}{A vector or matrix containing the values of the green band.}
}
\value{
A vector or matrix containing the calculated MCARI values.
}
\description{
The Modified Chlorophyll Absorption Ratio Index (MCARI) is a vegetation index that is calculated using the specified formula.
}
\examples{
red_band <- c(0.1, 0.2, 0.3, 0.4, 0.5)
redEdge_band <- c(0.15, 0.25, 0.35, 0.45, 0.55)
green_band <- c(0.2, 0.3, 0.4, 0.5, 0.6)
mcari(red_band, redEdge_band, green_band)
}
\references{
Daughtry, C. S., Walthall, C. L., Kim, M. S., De Colstoun, E. B., & McMurtrey Iii, J. E. (2000). Estimating corn leaf chlorophyll concentration from leaf and canopy reflectance. Remote sensing of Environment, 74(2), 229-239.
}
