% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndwi.R
\name{ndwi}
\alias{ndwi}
\title{Calculate the Normalized Difference Water Index (NDWI).}
\usage{
ndwi(green_band, nir_band)
}
\arguments{
\item{green_band}{A vector or matrix containing the values of the green band.}

\item{nir_band}{A vector or matrix containing the values of the NIR band.}
}
\value{
A vector or matrix containing the calculated NDWI values.
}
\description{
The Normalized Difference Water Index (NDWI) is a water index that is calculated using the specified formula: (green - NIR) / (green + NIR).
}
\examples{
green_band <- c(0.1, 0.2, 0.3, 0.4, 0.5)
nir_band <- c(0.2, 0.3, 0.4, 0.5, 0.6)
ndwi(green_band, nir_band)
}
\references{
McFeeters, S. K. (1996). The use of the Normalized Difference Water Index (NDWI) in the delineation of open water features. International journal of remote sensing, 17(7), 1425-1432.
}
