% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvi.R
\name{rvi}
\alias{rvi}
\title{Calculate the Ratio Vegetation Index.}
\usage{
rvi(red_band, nir_band)
}
\arguments{
\item{red_band}{A vector or matrix containing the values of the red band.}

\item{nir_band}{A vector or matrix containing the values of the NIR band.}
}
\value{
A vector or matrix containing the calculated Ratio Index values.
}
\description{
The Ratio Vegetation Index (RVI) is a vegetation index that is calculated using the formula: RVI = red / nir.
}
\examples{
red_band <- c(0.1, 0.2, 0.3, 0.4, 0.5)
nir_band <- c(0.2, 0.3, 0.4, 0.5, 0.6)
rvi(red_band, nir_band)
}
\references{
Jackson, R. D., & Huete, A. R. (1991). Interpreting vegetation indices. Preventive veterinary medicine, 11(3-4), 185-200.
}
