% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savi.R
\name{savi}
\alias{savi}
\title{Calculate the Soil-Adjusted Vegetation Index (SAVI).}
\usage{
savi(red_band, nir_band, L_savi)
}
\arguments{
\item{red_band}{A vector or matrix containing the red band values.}

\item{nir_band}{A vector or matrix containing the near-infrared (NIR) band values.}

\item{L_savi}{A numeric value representing the soil-adjustment factor for SAVI. Typically, L_savi is set to 0.5 for minimal soil influence.}
}
\value{
A vector or matrix containing the calculated SAVI values.
}
\description{
The Soil-Adjusted Vegetation Index (SAVI) is a vegetation index used in remote sensing to assess plant health while considering the influence of bare soil. It includes a soil-adjustment factor "L."
}
\examples{
red_band <- c(0.1, 0.2, 0.3, 0.4, 0.5)
nir_band <- c(0.2, 0.3, 0.4, 0.5, 0.6)
L_savi <- 0.5
savi(red_band, nir_band, L_savi)
}
\references{
Huete, A. R. (1988). A soil-adjusted vegetation index (SAVI). Remote sensing of environment, 25(3), 295-309.
}
