% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.R
\name{sr}
\alias{sr}
\title{Calculate the Simple Ratio Vegetation Index (SR).}
\usage{
sr(red_band, nir_band)
}
\arguments{
\item{red_band}{A vector or matrix containing the values of the red band.}

\item{nir_band}{A vector or matrix containing the values of the NIR band.}
}
\value{
A vector or matrix containing the calculated SR values.
}
\description{
The Simple Ratio Vegetation Index (SR) is a vegetation index that is calculated using the formula: SR = nir / red.
}
\examples{
red_band <- c(0.1, 0.2, 0.3, 0.4, 0.5)
nir_band <- c(0.2, 0.3, 0.4, 0.5, 0.6)
sr(red_band, nir_band)
}
\references{
Birth, G. S., & McVey, G. R. (1968). Measuring the color of growing turf with a reflectance spectrophotometer 1. Agronomy Journal, 60(6), 640-643.
}
